requirements.txt
setuptools==57.0.0
aiohttp==3.7.4.post0
altgraph==0.17
anyjson==0.3.3
asgi-redis==1.4.3
asgiref==3.3.4
astroid==2.5.6
async-timeout==3.0.1
attrs==21.2.0
auto-py-to-exe==2.9.0
autobahn==21.2.1
Automat==20.2.0
autopep8==1.5.7
boto3==1.17.71
botocore==1.20.71
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.3
django-allauth==0.44.0
django-cors-headers==3.7.0
django-countries==7.2.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-heroku==0.3.1
django-model-utils==4.1.1
django-pagination==1.0.7
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-safedelete==1.0.0
django-storages==1.11.1
django-widget-tweaks==1.4.8
djangochannelsrestframework==0.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-queryfields==1.0.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
Eel==0.12.4
Faker==8.1.4
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.0.1
incremental==21.3.0
inflection==0.5.1
install==1.3.4
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
macholib==1.14
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack-python==0.5.6
multidict==5.1.0
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
pep8==1.7.1
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyJWT==1.7.1
pylint==2.8.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
redis==2.10.6
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
service-identity==21.1.0
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
stripe==2.56.0
text-unidecode==1.3
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
whichcraft==0.6.1
whitenoise==5.2.0
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
